pyproject.toml
[tool.poetry]
name = "landmark-detection-cleardusk"
version = "0.1.0"
description = ""
authors = ["cleardusk <cleardusk@cleardusk.com>"]
license = "MIT"
readme = "README.md"
packages = [
    {include = "landmark_detection_cleardusk"}
]

[tool.poetry.dependencies]
sim3dr_cython = {path = "lib/sim3dr_cython", develop = true}
cpu_nms_cleardusk = {path = "lib/cpu_nms_cleardusk", develop = true}
python = ">=3.7, <3.8"
torch = "1.12.1"
torchvision = "0.13.1"
matplotlib = "3.5.3"
numpy = "1.21.6"
opencv-python = "4.6.0.66"
imageio = "2.22.1"
imageio-ffmpeg = "0.4.7"
pyyaml = "6.0"
tqdm = "4.64.1"
argparse = "1.4.0"
scikit-image = "0.19.3"
scipy = "1.7.3"
onnxruntime = "1.12.1"
gradio = "3.5"


[build-system]
requires = ["cython", "numpy", "poetry-core"]
build-backend = "poetry.core.masonry.api"
